requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
plotly>=5.15.0
Pillow>=10.0.0
opencv-python>=4.8.0

# Configuration and logging
omegaconf>=2.3.0
pyyaml>=6.0
tensorboard>=2.13.0
wandb>=0.15.0

# Data processing
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization and demo
gradio>=3.40.0
streamlit>=1.25.0

# Development tools
black>=23.0.0
ruff>=0.0.280
pytest>=7.4.0
pre-commit>=3.3.0

# Optional advanced dependencies
transformers>=4.30.0
timm>=0.9.0
kornia>=0.7.0
albumentations>=1.3.0
